+++ clusterability/R/clusterability.R
@@ -1,26 +1,26 @@
 "_PACKAGE"
 #' clusterability: a package to perform tests of clusterability
-#'
 # Copyright (C) 2026  Zachariah Neville, Naomi Brownstein, Andreas Adolfsson, Margareta Ackerman
-#
+
 # This program is free software; you can redistribute it and/or
 # modify it under the terms of the GNU General Public License
 # as published by the Free Software Foundation; either version 2
 # of the License, or (at your option) any later version.
-#
+
 # This program is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
 # MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 # GNU General Public License for more details.
-#
+
 # You should have received a copy of the GNU General Public License
 # along with this program; if not, write to the Free Software
 # Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
-#
-#' The \code{\link{clusterabilitytest}} function can test for
+
+#' @description
+#' The \code{\link{clusterabilitytest}} function tests for
 #' clusterability of a dataset, and the \code{\link[=print.clusterability]{print}} function
-#' to display output in the console. Below we include code to use with the provided example
-#' datasets. Please see the \code{clusterabilitytest} function for documentation on
+#' displays output in the console. Below, code is included to use with the provided example
+#' datasets. Please see the \code{clusterabilitytest()} function for documentation on
 #' available parameters.
 #'
 #' @examples