+++ clusterability/NEWS.md
@@ -1,3 +1,9 @@
+# clusterability 0.2.2.0
+* Code and accompanying documentation have been updated to improve readability and follow a more consistent style.
+
+# clusterability 0.2.1.0
+* Sparse PCA can now be used as a dimension reduction method. Both the elasticnet implementation and variable projection implementation are available.
+
 # clusterability 0.1.1.0
 * Silverman's Test no longer rounds p-values below 0.005 to 0
 * Clean up code for examples and plots